pyproject.toml
[build-system]
requires      = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name            = "edge-llm"
version         = "1.0.0"
description     = "Unified, lightweight OpenAI-compatible API server for MLX and GGUF models."
authors         = [{ name = "Fred Bliss" }]
readme          = "README.md"
license         = { file = "LICENSE" }
requires-python = ">=3.11"
classifiers     = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: MacOS",
  "Operating System :: POSIX :: Linux",
]

# -------------------------------------------------------------------
# Core runtime deps – array of PEP 508 strings (required by spec)
# -------------------------------------------------------------------
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pydantic",
  "python-multipart",
  "pyyaml",
  "requests",
  "rich",
  "numpy",
  "Pillow",
  "mlx-lm",
  "llama-cpp-python",
]

# ─── Optional groups ────────────────────────────────────────────────────────
[project.optional-dependencies]
test = ["pytest>=8", "pytest-mock>=3"]

[project.scripts]
edge-llm = "edge_llm.server:main"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"edge_llm.templates" = ["*.jinja2", "*.tpl", "*.yaml"]
